move the scripts folder
